Add PHPUnit test suite for theme helper functions
Set up a lightweight unit-testing harness (Composer + PHPUnit + Brain
Monkey) so the theme's business logic can be tested without a running
WordPress install. WordPress functions are mocked, so the suite stays
fast and DB-free.

Covers the pure helpers in inc/template-functions.php:
- translation lookup and language detection (bg/en, fallbacks)
- promotion logic: active/expired checks, price and percentage calc
- exhibition status and date-range formatting
- publication date/meta assembly
- testimonial star rendering and author meta

Fix sculpture_get_promotion_percentage() returning a float from its
manual-price branch while the docblock and the percentage branch return
an int — surfaced by the new tests. Now consistently returns an int.

+++ inc/template-functions.php
@@ -1088,7 +1088,7 @@ function sculpture_get_promotion_percentage($post_id = null)
     $manual_price = get_field("promotion_price", $post_id);
 
     if ($manual_price && $price) {
-        return round((1 - $manual_price / $price) * 100);
+        return (int) round((1 - $manual_price / $price) * 100);
     }
 
     return $percentage ? intval($percentage) : null;
